Make thiran allpass use its templated type for processing
Also, correctly declare the fractional delay filters in
plugin processor to use the type given to them in processing.

+++ src/PluginProcessor.cpp
@@ -558,7 +558,7 @@ void ValentineAudioProcessor::initializeDSP()
     bitCrush = std::make_unique<Bitcrusher>();
 
     for (auto& filter : fracDelayFilters)
-        filter.reset (new FirstOrderThiranAllpass<double>);
+        filter.reset (new FirstOrderThiranAllpass<float>);
 
     for (auto& buffer : unProcBuffers)
     {
